serviceBusTestingWithContainer/requirements.txt
azure-functions
pillow
pdf2image
pytesseract
azure-storage-blob
asyncio
aiohttp
